### 2023-12-09 #6134 @editorialbot
[PRE REVIEW]: Augmenty: A Python Library for Structured Text Augmentation
**Submitting author:** @KennethEnevoldsen (<a href="http://orcid.org/0000-0001-8733-0966">Kenneth C. Enevoldsen</a>)
**Repository:** https://github.com/KennethEnevoldsen/augmenty
**Branch with paper.md** (empty if default branch):
**Version:** 1.4.2
