### 2023-12-18 96fff7c
paper: Updated based on feedback from @HLasse
