### 2023-12-26 da97b19
fix bug in subset
