### 2023-12-26 9027a49
fix eos bounds when creating paragraph subsets
