### 2023-12-26 0e4b222
fix: cleaning up unn. type ignores
