+++ tests/test_doc.py
@@ -74,8 +74,16 @@ def test_paragraph_subset_augmenter(nlp_en: Language):
     assert aug_docs[0].text == ""  # type: ignore
 
 
-def test_paragraph_subset_augmenter_edge_cases(nlp_en: Language):
-    random.seed(1)
+def test_paragraph_subset_augmenter_issue_when_ent_at_sentence_bound(nlp_en: Language):
+    """
+    There were issues with the augmenter when the entity was at the sentence boundery.
+
+    Where it e.g. would return "I am John Heinz", even though the entity was "John" and "Heinz" with a sentence boundery in between.
+    """
+    p_subset_aug = create_paragraph_subset_augmenter_v1(
+        min_paragraph=1,
+        max_paragraph=0.50,
+    )
 
     doc = Doc(
         nlp_en.vocab,
@@ -83,17 +91,10 @@ def test_paragraph_subset_augmenter_edge_cases(nlp_en: Language):
         ents=["O", "O", "B-PER", "B-PER", "O", "O", "O"],
         sent_starts=[True, False, False, True, False, False, False],
     )
-
-    p_subset_aug = create_paragraph_subset_augmenter_v1(
-        min_paragraph=1,
-        max_paragraph=0.50,
-    )
-
+    random.seed(1)
     aug_docs = list(augmenty.docs([doc], p_subset_aug, nlp_en))
     doc = aug_docs[0]
     assert doc.text == "I am John " or doc.text == "Heinz is happy . "
-
-    random.seed(42)
 
     doc = Doc(
         nlp_en.vocab,
@@ -101,6 +102,7 @@ def test_paragraph_subset_augmenter_edge_cases(nlp_en: Language):
         ents=["O", "O", "B-PER", "I-PER", "B-PER", "O", "O", "O"],
         sent_starts=[True, False, False, False, True, False, False, False],
     )
+    random.seed(42)
     aug_docs = list(augmenty.docs([doc], p_subset_aug, nlp_en))
     doc = aug_docs[0]
     assert doc.text == "I am John Doe " or doc.text == "Heinz is happy . "

[…]
