### 2023-12-26 4672b01
fix: Added missing type ignores
