### 2024-01-22 #6134 @KennethEnevoldsen
Here is 5 people from the reviewer list. Generally, I would go for people with an NLP/Python background. Augmentation is not as popular in NLP as in computer vision so at least one reviewer with experience in CV might be ideal, though not required (as they most likely also will have some experience in augmentation).

- [dmnapolitano](https://github.com/dmnapolitano)
- [harishsiravuri](https://github.com/harishsiravuri)
- [jhagerer](https://github.com/jhagerer)
- [Gigi1111](https://github.com/Gigi1111)
- [jinhangjiang](https://github.com/jinhangjiang)
