### 2024-01-22 #6134 @arfon
:wave: @dmnapolitano @harishsiravuri @jinhangjiang @Gigi1111 would any of you be willing to review this submission for JOSS? The submission under consideration is _Augmenty: A Python Library for Structured Text Augmentation_

Many thanks
Arfon
