### 2024-01-31 #6134 @arfon
@sap218 @wdduncan @balhoff @jhagerer – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _Augmenty: A Python Library for Structured Text Augmentation_

Many thanks
Arfon
