### 2024-03-10 #6370 @sap218
Hi all! Happy to have started this review - looking great and I've successfully ran the "Simple Example" from the repo.

Whilst I am looking into the code, I have some initial comments about the Paper:

- Needs a little introduction to text augmentation itself - e.g. "Text augmentation is creating pseudo/AI-generated text" etc. just to give the reader a little more context
- Fix references: spaCy [honnibal_efficient_2020] line 12/41, & NLTK [bird2009natural] line 64 & ref on line 66
- On line 22 you have "natural language processing (NLP)" but you can move this to line 7 and instead use the abbreviation NLP from that point onward
- On line 21 "used for both training more robust models and for evaluating the ability of the models to handle pertubations" this is actually repeated from line 6
- Moreover, I would like if Summary to focus more on Augmenty and keep Statement of Need about other packages
- I would like Statement of Need to end on the introduction to Augmenty (so it flows nicely to Features and Functions) and so move the final paragraph (lines 46-52) to before the introduction of Augmenty

Some little things I wanted to mention - I appreciate that the author acknowledges the contributors & the documentation is very good and thorough!

Happy to check these things off when addressed - please let me know if you have any questions (hope it all makes sense)
Will continue with the review this week! :)
