### 2024-03-11 4bf0cb2
docs: Updated paper with feedback from @sap218
