### 2024-03-11 #6370 @KennethEnevoldsen
Thanks for the suggestions @sap218 I have fixed 1, 2, 3, 6.

re. 4) it is a summary so some degree of duplication is expected. If you feel it is a big problem I will reformulate the sentence, but I don't see it as problematic.

re. 5) I believe the summary should include both the relevant context as well as the package. If you have specific information you would like added to the summary I would be more than happy to rewrite it. Statement of needs is structured such that existing augmentation libraries are contrasted with augmenty, I believe this is the best way to highlight both the need for the package, but also what it tried to solve and what it does not.
