### 2024-03-11 46220c4
docs: fix minor spelling mistake
