+++ paper/paper.md
@@ -32,7 +32,7 @@ Existing tools such as `textgenie` [@pandya_hetpandyatextgenie_2023], and `texta
 Other tools for data augmentation focus on specific downstream application such as `textattack` [@morris2020textattack] which is useful for adversarial attacks of classification systems or `robustnessgym` [@goel-etal-2021-robustness] which is useful for evaluating robustness of classification systems. 
 
 
-`Augmenty` introduced a flexible and easy-to-use interface for structured text augmentation, seeking to augment the annotations along with the text. `Augmenty` is built to integrate well with `spaCy` [@spacy] and seeks to be compatible with the broad set of tasks supported by `spaCy`. Augmenty provides augmenters which take a spaCy `Doc`-object (but works just as well with `string`-objects) and return a new `Doc`-object with the augmentations applied. This allows for augmentations of both the text and the annotations present in the `Doc`-object.
+`Augmenty` introduces a flexible and easy-to-use interface for structured text augmentation, seeking to augment the annotations along with the text. `Augmenty` is built to integrate well with `spaCy` [@spacy] and seeks to be compatible with the broad set of tasks supported by `spaCy`. Augmenty provides augmenters which take a spaCy `Doc`-object (but works just as well with `string`-objects) and return a new `Doc`-object with the augmentations applied. This allows for augmentations of both the text and the annotations present in the `Doc`-object.
 `Augmenty` does not seek to replace useful tools such as `textattack`, but seeks to provide a general purpose tool for augmentation of both the text and its annotations. This allows for augmentations within a range of applications such as named entity recognition, part-of-speech tagging, and dependency parsing.
 
 
