### 2024-03-11 #6370 @sap218
1. In Documentation "Installation instructions: Is there a clearly-stated list of dependencies?" I might be missing this but I can't find these dependencies easily?
2. In Documentation "Functionality documentation" I am happy with your documentation! This isn't necessary to change but the table on page "Overview of Augmenters" is a little difficult to navigate for me - perhaps font is too big or you could move description into same column as Augmenter name - but again you don't need to change this if you wish to keep it as is :)
