### 2024-03-18 #6370 @wdduncan
I've started reviewing. So far here is what I've done using `python 3.11.8` and `pip` version `24.0`. I am running code on an a MacBook Pro with an Apple M1 Max processor and 64GB of RAM; OSX version Sonoma 14.3.1.

* created a virtual environment with `python -m venv`
* installed using `pip install augmenty`
  * when I try to install using `pip install augmenty[all]`, it fails with `zsh: no matches found: augmenty[all]`. any ideas why?
* run `make test` -> this **fails** with error `make: *** [test] Abort trap: 6`. perhaps a memory error?
* it would helpful if the instructions included instructions about downloading space models; e.g. I had run `python -m spacy download en_core_web_sm` to get `en_core_web_sm`. also, I am not an expert using `spacy`, but to get your github code to run, I had to load  `en_core_web_sm` using:
```python
import en_core_web_sm
nlp = en_core_web_sm.load()
```
* when I run:
``python
for doc in augmenty.docs(docs, augmenter=entity_augmenter, nlp=nlp):
    print(doc)
```
it fails with error `NameError: name 'docs' is not defined`.

Am I using the correct version of python?
