### 2024-03-19 #6370 @KennethEnevoldsen
Hi @sap218

1) the [pyproject.toml](https://github.com/KennethEnevoldsen/augmenty/blob/main/pyproject.toml) specified all the required dependencies. The pyproject.toml is a more recent alternative to setup.py and setup.cfg (I believe it is the recommended format)

2) I agree the table could be better. Currently, there is sadly not a lot of support for tables in the sphinx ecosystem. I will see if I can make it clearer though.
