+++ docs/conf.py
@@ -112,4 +112,3 @@ html_theme_options = {
     "sidebar_hide_name": True,
     "navigation_with_keys": True,
 }
-
