### 2024-03-19 #6370 @KennethEnevoldsen
Hi @wdduncan

1) When running on zsh I believe you need to write `pip install "augmenty[all]"` (I will change to documentation to match)
2) to run the `make test` you will have to install it using `make install`, this installs the development dependencies as well. I have tried to make this more clear in the contributing.md.
3) I agree that it should be clear that it needs to be downloaded. I have added a comment:
```python
import spacy
import augmenty

nlp = spacy.load("en_core_web_md")
# if not installed run: python -m spacy download en_core_web_md

docs = nlp.pipe(["Augmenty is a great tool for text augmentation"])

entity_augmenter = augmenty.load("ents_replace_v1",
                                 ent_dict = {"ORG": [["spaCy"], ["spaCy", "Universe"]]}, level=1)

for doc in augmenty.docs(docs, augmenter=entity_augmenter, nlp=nlp):
    print(doc)
```

3.1) You shouldn't need to use this syntax. However, that seems to be a spacy issue. If you only have issues with it when installing augmenty let me know. (I suspect it might be spacy assuming something about your paths)

4) I believe you need to run the following line:
```
docs = nlp.pipe(["Augmenty is a great tool for text augmentation"])
```

If this step failed it could suggest an error in your spacy pipeline installation.
