### 2024-03-19 477895c
fix: updated tests dependencies due to issues with pytest-fixtures:
https://github.com/TvoroG/pytest-lazy-fixture/issues/65
