+++ makefile
@@ -1,6 +1,7 @@
 install:
 	@echo "--- 🚀 Installing project ---"
 	pip install -e ".[dev, docs, tests,tutorials,all,da]" 
+	pip install -r tests/requirements.txt
 
 static-type-check:
 	@echo "--- 🔍 Running static type check ---"
