+++ tests/test_all_augmenters.py
@@ -1,6 +1,5 @@
 """Pytest script for testing all augmenters in a variety of cases."""
 
-
 from typing import Iterable
 
 import augmenty
@@ -80,9 +79,7 @@ augmenters_args = {
 }
 
 
-@pytest.mark.parametrize(
-    "aug,args", [(k, augmenters_args[k]) for k in augmenters_args]
-)  # noqa
+@pytest.mark.parametrize("aug,args", [(k, augmenters_args[k]) for k in augmenters_args])  # noqa
 @pytest.mark.parametrize("level", [0.1, 0.5, 1])
 @pytest.mark.timeout(100)
 @pytest.mark.parametrize(
