### 2024-03-20 #6370 @sap218
Hi @KennethEnevoldsen

1. Thanks for letting me know! I've ticked off this item (Installation instructions: Is there a clearly-stated list of dependencies?)
2. I appreciate this - no worries if you can't come to a solution as I do think the table does the job well 😃

I should be completed with my review at the end of this week/next week (functionality and automated tests are left)
