### 2024-03-25 #6370 @wdduncan
@KennethEnevoldsen I ran the command `pip install "augmenty[all]"`. This worked fine.

When I run
```python
nlp = spacy.load("en_core_web_md")
docs = nlp.pipe(["Augmenty is a great tool for text augmentation"])
entity_augmenter = augmenty.load("ents_replace_v1", ent_dict = {"ORG": [["spaCy"], ["spaCy", "Universe"]]}, level=1)

for doc in augmenty.docs(docs, augmenter=entity_augmenter, nlp=nlp):
    print(doc)
```
it returns
```python
Augmenty is a great tool for text augmentation
```
Should it return?:
```python
spaCy Universe is a great tool for text augmentation.
```
