### 2024-03-27 #6370 @sap218
@KennethEnevoldsen I have tested functionality and all seems good! I can sign off the final checklist item Software Paper: References once the reference on line 67 [pennington] is fixed 😄 thank you for your patience!
