### 2024-03-27 #6370 @sap218
Just a note @KennethEnevoldsen I encountered a small `setuptools` problem w/ ubuntu (https://github.com/pypa/setuptools/issues/3269) when trying to run the tests (the fix was here https://pip.pypa.io/en/stable/installation/#get-pip-py). This isn't a problem w/ your package at all but wanted to let you know in case you wanted to update documentation...

BUT, I did run into a test issue - when I ran `python -m pip install -e ".[tests]"` this doesn't include dacy and so I couldn't run the tests... to resolve this I had to also run `python -m pip install -e ".[da]"` to finally be able to run `python -m pytest` - documentation should reflect this - or a fix to run the tests without dacy
