### 2024-04-01 #6370 @KennethEnevoldsen
@sap218 thanks for letting me know. Sounds like a very tricky problem, that is good to know about!

> BUT, I did run into a test issue - when I ran python -m pip install -e ".[tests]" this doesn't include dacy and so I couldn't run the tests... to resolve this I had to also run python -m pip install -e ".[da]" to finally be able to run python -m pytest - documentation should reflect this - or a fix to run the tests without dacy

I have fixed that. It is actually stated in the contributing guidelines (but wrongly stated in the FAQ). I have removed the FAQ section on running tests so there is no duplicate information.

@wdduncan thanks for catching the mistake. This seems to be caused the by fact that "ORG" is replace with "GPE" (geopolitical entity) in the latest spaCy models. You should be able to fix it using:

```python
nlp = spacy.load("en_core_web_md")
docs = nlp.pipe(["Augmenty is a great tool for text augmentation"])
entity_augmenter = augmenty.load("ents_replace_v1", ent_dict = {"GPE": [["spaCy"], ["spaCy", "Universe"]]}, level=1)

for doc in augmenty.docs(docs, augmenter=entity_augmenter, nlp=nlp):
    print(doc)
# spaCy Universe is a great tool for text augmentation.
```

I have updated the documentation to match
