+++ docs/faq.rst
@@ -8,36 +8,6 @@ If you use this library in your research, it would be much appreciated it if you
 "cite this repository" on the `github page <https://github.com/KennethEnevoldsen/augmenty>`__ for an up to date citation.
 
 
-How do I test the code and run the test suite?
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-This package comes with an extensive test suite. In order to run the tests,
-you'll usually want to clone the repository and build the package from the
-source. This will also install the required development dependencies
-and test utilities defined in the extras_require section of the :code:`pyproject.toml`.
-
-.. code-block:: bash
-
-   pip install -e ".[tests]"
-
-   python -m pytest
-
-
-which will run all the test in the `tests` folder.
-
-Specific tests can be run using:
-
-.. code-block:: bash
-
-   python -m pytest tests/desired_test.py
-
-If you want to check code coverage you can run the following:
-
-.. code-block::
-
-   python -m pytest --cov=.
-
-
 Does this package run on X?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
