### 2024-04-01 #6370 @KennethEnevoldsen
@sap218 I have added a link to the table in the readme to the CONTRIBUTING.md file

Oh right so the issue, might be very dependent on what version of spacy pipelines is being used (one might be using a newer or an older pipeline). As the first example is important I have added a few more details. Hopefully these should allow the user to better know what is going on. Feel free to say if it doesn't help.

```python
import augmenty
import spacy

nlp = spacy.load("en_core_web_md")
# if not installed run: python -m spacy download en_core_web_md

doc = nlp("Augmenty is a great tool for text augmentation")

# check that the pipeline detects the entities (this done by SpaCy and is not a 100%)
print(doc.ents)
# (Augmenty,) is detected as an entity

doc.ents[0].label_
# 'GPE'. Depending on the model, the label might be different (e.g. 'ORG')

entity_augmenter = augmenty.load(
    "ents_replace_v1", ent_dict={"GPE": [["spaCy"], ["spaCy", "Universe"]]}, # label=GPE,
    level=1
)

for augmented_doc in augmenty.docs([doc], augmenter=entity_augmenter, nlp=nlp):
    print(augmented_doc)

```

> Also interestingly, GPE works instead of the NER tag LOC (location) but not sure if this is intention?

My guess is that New York is considered a GPE and not a location. This depends on the annotation guidelines used on the dataset, however generally a location have no agency and e.g. can't have impose rules ("New York [GPE] imposes a new rule requiring citizens to ..."), while a GPE can have agency. Some times it is also annotated depending on the context, i.e. "I was in New York [LOC] during the week" (no agency).
