### 2024-04-01 3eb9ad0
docs: updated main example
