### 2024-04-03 #6370 @wdduncan
> > @arfon I ran the @editorialbot generate my checklist command. See [#6370 (comment)](https://github.com/openjournals/joss-reviews/issues/6370#issuecomment-2018565334). Where is the checklist generated?
>
> Weird, it should have edited [this comment](https://github.com/openjournals/joss-reviews/issues/6370#issuecomment-2018565334), replacing it with the checklist. @wdduncan – would you mind trying again please?

Now it worked :)
