### 2024-04-20 f7339cc
docs: Update paper.md
