### 2024-04-25 #6370 @crvernon
:wave: @KennethEnevoldsen - one last thing before we accept...

Could you edit your metadata for the license stated in this archive https://zenodo.org/records/11002422 to match what is used by your package?

No need to do another release, you can just edit the existing record on Zenodo.

Let me know when this is done and I'll move forward with accepting the submission for publications.  Thanks!
