### 2024-04-27 #6370 @crvernon
Congratulations on your new publication @KennethEnevoldsen! Many thanks to @arfon and to reviewers @sap218 and @wdduncan for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
