### 2024-04-29 #6370 @KennethEnevoldsen
Thanks @crvernon, @arfon and the reviewers @wdduncan and @sap218 - very happy about the feedback always a good experience to publish with joss
